Add useQueryString
Go on #36
